Move Square's tests to the paths their modules mirror
The mutation gate finds a source's direct tests by the mirror path alone, so
a source whose test sits under another name has no direct suite and the gate
refuses to run for it. Square's tests were named for what they do rather than
for the module they cover.

Each file now sits at its module's mirror path. The connection checks move out
of the client suite into their own file, because they cover connection.ts.
