+++ test/shared/square/checkout/link.test.ts
