+++ test/shared/square/checkout/validation.test.ts
