+++ test/shared/square/client/rest.test.ts
