+++ test/shared/square/payment-outcomes/read-money.test.ts
